Bump version to 0.0.8, add crop region to filter data, and update build configurations
