+++ CMakeLists.txt
@@ -22,7 +22,7 @@ if(ENABLE_FRONTEND_API)
 endif()
 
 if(ENABLE_QT)
-  find_qt(COMPONENTS Widgets Core)
+  find_package(Qt6 COMPONENTS Widgets Core)
   target_link_libraries(${CMAKE_PROJECT_NAME} PRIVATE Qt::Core Qt::Widgets)
   target_compile_options(
     ${CMAKE_PROJECT_NAME} PRIVATE $<$<C_COMPILER_ID:Clang,AppleClang>:-Wno-quoted-include-in-framework-header
@@ -78,7 +78,14 @@ endif()
 include(cmake/BuildInja.cmake)
 target_link_libraries(${CMAKE_PROJECT_NAME} PRIVATE inja)
 
-target_sources(${CMAKE_PROJECT_NAME} PRIVATE src/plugin-main.c src/obs-utils.cpp src/tesseract-ocr-utils.cpp
-                                             src/ocr-filter.cpp src/ocr-filter-info.c src/text-render-helper.cpp)
+target_sources(
+  ${CMAKE_PROJECT_NAME}
+  PRIVATE src/ocr-filter-properties.cpp
+          src/plugin-main.c
+          src/obs-utils.cpp
+          src/tesseract-ocr-utils.cpp
+          src/ocr-filter.cpp
+          src/ocr-filter-info.c
+          src/text-render-helper.cpp)
 
 set_target_properties_plugin(${CMAKE_PROJECT_NAME} PROPERTIES OUTPUT_NAME ${_name})
