+++ cmake/common/buildnumber.cmake
@@ -12,8 +12,14 @@ if(NOT DEFINED PLUGIN_BUILD_NUMBER AND EXISTS "${_BUILD_NUMBER_CACHE}")
   file(READ "${_BUILD_NUMBER_CACHE}" PLUGIN_BUILD_NUMBER)
   math(EXPR PLUGIN_BUILD_NUMBER "${PLUGIN_BUILD_NUMBER}+1")
 elseif(NOT DEFINED PLUGIN_BUILD_NUMBER)
-  if($ENV{CI} AND $ENV{GITHUB_RUN_ID})
-    set(PLUGIN_BUILD_NUMBER "$ENV{GITHUB_RUN_ID}")
+  if($ENV{CI})
+    if($ENV{GITHUB_RUN_ID})
+      set(PLUGIN_BUILD_NUMBER "$ENV{GITHUB_RUN_ID}")
+    elseif($ENV{GITLAB_RUN_ID})
+      set(PLUGIN_BUILD_NUMBER "$ENV{GITLAB_RUN_ID}")
+    else()
+      set(PLUGIN_BUILD_NUMBER "1")
+    endif()
   else()
     set(PLUGIN_BUILD_NUMBER "1")
   endif()
