+++ cmake/windows/resources/installer-Windows.iss.in
@@ -20,9 +20,6 @@ OutputBaseFilename={#MyAppName}-{#MyAppVersion}-Windows-Installer
 Compression=lzma
 SolidCompression=yes
 DirExistsWarning=no
-VersionInfoVersion={#MyAppVersion}
-AppCopyright=Copyright (C) 2023 {#MyAppPublisher}
-DisableProgramGroupPage=Yes
 
 [Languages]
 Name: "english"; MessagesFile: "compiler:Default.isl"
