+++ data/locale/en-US.ini
@@ -1,5 +1,5 @@
 OCRFilter="OCR"
-PageSegmentationMode="Page Segmentation Mode"
+PageSegmentationMode="Segmentation Mode"
 Language="Language"
 EnableSmoothing="Enable Smoothing"
 ConfThreshold="Confidence Threshold"
@@ -20,7 +20,7 @@ OutputFilePath="Output File Path"
 BinarizationMode="Binarization Mode"
 BinarizationThreshold="Binarization Threshold"
 BinarizationBlockSize="Binarization Block Size"
-PreviewBinarization="Preview Binarization"
+PreviewBinarization="Preview OCR Image"
 RescaleImage="Rescale Image"
 RescaleTargetSize="Rescale Target Size"
 DilationIterations="Dilation Iterations"
@@ -29,7 +29,7 @@ DetectionBoxesMask="Detection Boxes Mask"
 TextRendering="Text Overlay"
 TextWithBackground="Text with Background"
 SelectPresets="Select Preset"
-CharWhitelistPreset="Character Whitelist Preset"
+CharWhitelistPreset="Whitelist Preset"
 NumericPunctuation="Numeric / Punctuation"
 OutputFlatten="Flatten Output to Single Line"
 OutputFileAppend="Append to File?"
